Add takeUntil and edited other pull request issues
