+++ src/app/app.component.scss
@@ -1,3 +1,9 @@
+@import "shared/style/variables";
+
 main {
   padding: 5px;
+}
+
+.error {
+  color: $colorRedForError;
 }
