--- src/app/shared/style/style.scss
@@ -1,5 +1,0 @@
-@import "variables";
-
-.error {
-  color: $colorRedForError;
-}
