+++ src/app/shelters/models/address-shelter.interface.ts
@@ -1,0 +1,7 @@
+export interface AddressShelter {
+  country: string;
+  region: string;
+  city: string;
+  street: string;
+  house: number;
+}
