+++ src/app/shelters/models/shelter.interface.ts
@@ -7,11 +7,3 @@ export interface Shelter {
   name: string;
   photoPath: string;
 }
-
-export interface AddressShelter {
-  country: string;
-  region: string;
-  city: string;
-  street: string;
-  house: number;
-}
